Update gems to satisfy bundler
Having decided to finally embark on the journey to upgrade this site so
that it is running on recent versions of Ruby, Rails and gems, the first
step was to enable `bundle install` to succeed. Whilst the first error
was a failure to install the scrypt gem, which was fixed by upgrading
authlogic, the acts-as-taggable-on gem had to also be fixed because the
version specified in the Gemfile had been yanked.