Upgrade Ruby to 2.3.8
In the process of trying to get the rails server to start without
errors, I encountered an error that required reinstalling Ruby. So,
before doing so, I decided it was time to at least update to the latest
version in the 2.3 path. Hopefully this will be short-lived before
upgrading Ruby to something closer to 2.6.1!